Changed QReward type from int to double. Commented the simulation class. There are some problems with the joints going out of range.

+++ src/agent_learner.hpp
@@ -15,7 +15,7 @@ friend class test_case_name##_##test_name##_Test
 // The state and reward are calculated by the agent from the simulation's
 // response.
 typedef int QState;
-typedef int QReward;
+typedef double QReward;
 
 class AgentLearner {
 public:
